pyproject.toml
[project]
name = "s4-landmark-ir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.5.0",
    "faiss-cpu>=1.13.2",
    "ftfy>=6.3.1",
    "ipywidgets>=8.1.8",
    "loralib>=0.1.2",
    "matplotlib>=3.10.8",
    "openai>=2.20.0",
    "outlines>=1.2.10",
    "polars>=1.38.0",
    "qdrant-client>=1.16.2",
    "randomname>=0.2.1",
    "scipy>=1.17.0",
    "setuptools>=80.10.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
]

[dependency-groups]
dev = [
    "iprogress>=0.4",
    "ipykernel>=7.1.0",
]
